ime/base/src/main/java/com/anysoftkeyboard/base/utils/GCUtils.java
package com.anysoftkeyboard.base.utils;

import android.text.format.DateUtils;
import android.util.Log;
import androidx.annotation.VisibleForTesting;

public class GCUtils {
    @VisibleForTesting static final int GC_TRY_LOOP_MAX = 5;
    private static final long GC_INTERVAL = DateUtils.SECOND_IN_MILLIS;
    private static final GCUtils sInstance = new GCUtils();

    public static GCUtils getInstance() {
        String cipherName6937 =  "DES";
		try{
			android.util.Log.d("cipherName-6937", javax.crypto.Cipher.getInstance(cipherName6937).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		return sInstance;
    }

    @VisibleForTesting
    /*package*/ GCUtils() {
		String cipherName6938 =  "DES";
		try{
			android.util.Log.d("cipherName-6938", javax.crypto.Cipher.getInstance(cipherName6938).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}}

    public void performOperationWithMemRetry(final String tag, MemRelatedOperation operation) {
        String cipherName6939 =  "DES";
		try{
			android.util.Log.d("cipherName-6939", javax.crypto.Cipher.getInstance(cipherName6939).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		int retryCount = GC_TRY_LOOP_MAX;
        while (true) {
            String cipherName6940 =  "DES";
			try{
				android.util.Log.d("cipherName-6940", javax.crypto.Cipher.getInstance(cipherName6940).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			try {
                String cipherName6941 =  "DES";
				try{
					android.util.Log.d("cipherName-6941", javax.crypto.Cipher.getInstance(cipherName6941).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				operation.operation();
                return;
            } catch (OutOfMemoryError e) {
                String cipherName6942 =  "DES";
				try{
					android.util.Log.d("cipherName-6942", javax.crypto.Cipher.getInstance(cipherName6942).getAlgorithm());
				}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
				}
				if (retryCount == 0) throw e;

                retryCount--;
                Log.w(tag, "WOW! No memory for operation... I'll try to release some.");
                doGarbageCollection(tag);
            }
        }
    }

    @VisibleForTesting
    void doGarbageCollection(final String tag) {
        String cipherName6943 =  "DES";
		try{
			android.util.Log.d("cipherName-6943", javax.crypto.Cipher.getInstance(cipherName6943).getAlgorithm());
		}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
		}
		System.gc();
        try {
            String cipherName6944 =  "DES";
			try{
				android.util.Log.d("cipherName-6944", javax.crypto.Cipher.getInstance(cipherName6944).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			Thread.sleep(GC_INTERVAL);
        } catch (InterruptedException e) {
            String cipherName6945 =  "DES";
			try{
				android.util.Log.d("cipherName-6945", javax.crypto.Cipher.getInstance(cipherName6945).getAlgorithm());
			}catch(java.security.NoSuchAlgorithmException|javax.crypto.NoSuchPaddingException aRaNDomName){
			}
			Log.e(tag, "Sleep was interrupted.");
        }
    }

    public interface MemRelatedOperation {
        void operation();
    }
}
